補充

+++ scene/game/game.gd
@@ -146,12 +146,17 @@ func to_continue():
 
 
 func set_progress(ps: float):
+	if progress_bar.value == ps:
+		return
+	
 	if pb_tween:
 		pb_tween.kill()
 	pb_tween = progress_bar.create_tween()
 	pb_tween.set_trans(Tween.TRANS_QUART)
 	pb_tween.set_ease(Tween.EASE_OUT)
 	pb_tween.tween_property(progress_bar, "value", ps, 1)
+	pb_tween.parallel().tween_property($progress, "scale", Vector2(1.05, 1.05), 1)
+	pb_tween.tween_property($progress, "scale", Vector2(1, 1), 1)
 
 
 func set_character_tween():
